@@ -3,7 +3,6 @@
 namespace App\Console;
 
 use App\Jobs\AnularProformaJob;
-use App\Jobs\Bodega\NotificarPedidoParcial;
 use App\Jobs\Bodega\NotificarPedidoParcialJob;
 use App\Jobs\MyJobExample;
 use App\Jobs\NotificarPermisoJob;
@@ -23,7 +22,7 @@ class Kernel extends ConsoleKernel
     /**
      * Define the application's command schedule.
      *
-     * @param  \Illuminate\Console\Scheduling\Schedule  $schedule
+     * @param Schedule $schedule
      * @return void
      */
     protected function schedule(Schedule $schedule)
