@@ -2,7 +2,7 @@
 
 namespace App\Http\Resources\Vehiculos;
 
-use App\Models\Vehiculos\ChecklistAccesoriosVehiculo;
+use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\JsonResource;
 use Src\Shared\Utils;
 
@@ -11,8 +11,8 @@ class BitacoraVehicularResource extends JsonResource
     /**
      * Transform the resource into an array.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @return array|\Illuminate\Contracts\Support\Arrayable|\JsonSerializable
+     * @param  Request  $request
+     * @return array
      */
     public function toArray($request)
     {
