@@ -2,6 +2,7 @@
 
 namespace App\Models\RecursosHumanos\SeleccionContratacion;
 
+use App\Models\Archivo;
 use App\Models\Autorizacion;
 use App\Models\Cargo;
 use App\Models\Empleado;
@@ -12,28 +13,17 @@ use Illuminate\Database\Eloquent\Model;
 use OwenIt\Auditing\Contracts\Auditable;
 use OwenIt\Auditing\Auditable as AuditableModel;
 
-class SolicitudPuestoEmpleo extends Model implements Auditable
+/**
+ * @method static create($validated)
+ * @method static ignoreRequest(string[] $array)
+ */
+class SolicitudPersonal extends Model implements Auditable
 {
     use HasFactory;
     use AuditableModel;
     use UppercaseValuesTrait;
     use Filterable;
-    protected $table = 'rrhh_solicitudes_puestos_empleos';
-    // 'nomina y roles de pago';  prefix=rrhh_nomina_
-    // 'seleccion y contratacion de personal'; prefix=rrhh_contratacion_
-    // 'control de asistencia'; prefix=rrhh_asistencia_
-    // 'otros submodulos'; prefix=rrhh_
-
-    /**
-     * Este comando sirve para hacer rollback y migrar en un solo paso
-     *  las ultimas migraciones hacia atras segun el valor pasado a step
-     * 
-     * php artisan migrate:refresh --step=3
-     * 
-     * `Nota` Si no se usa step se ejecutaran todas las migraciones desde cero
-     */
-
-
+    protected $table = 'rrhh_contratacion_solicitudes_nuevas_vacantes';
 
     protected $fillable = [
         'nombre',
@@ -43,14 +33,17 @@ class SolicitudPuestoEmpleo extends Model implements Auditable
         'autorizacion_id',
         'cargo_id',
         'anios_experiencia',
+        'areas_conocimiento',
         'descripcion',
 
     ];
-    private static $whiteListFilter = ['*'];
+    private static array $whiteListFilter = ['*'];
+
     public function tipoPuesto()
     {
-        return $this->hasOne(TipoPuestoTrabajo::class, 'id', 'tipo_puesto_id');
+        return $this->hasOne(TipoPuesto::class, 'id', 'tipo_puesto_id');
     }
+
     public function cargo()
     {
         return $this->hasOne(Cargo::class, 'id', 'cargo_id');
@@ -62,5 +55,18 @@ class SolicitudPuestoEmpleo extends Model implements Auditable
     public function autorizacion()
     {
         return $this->hasOne(Autorizacion::class, 'id', 'autorizacion_id');
+    }
+    public function formacionesAcademicas()
+    {
+        return $this->morphMany(FormacionAcademica::class, 'formacionable');
+    }
+
+    /**
+     * Relacion polimorfica con Archivos uno a muchos.
+     *
+     */
+    public function archivos()
+    {
+        return $this->morphMany(Archivo::class, 'archivable');
     }
 }
