@@ -6,13 +6,20 @@ use App\Models\ActividadRealizada;
 use App\Models\Empleado;
 use App\Models\Notificacion;
 use App\Traits\UppercaseValuesTrait;
+use Carbon\Carbon;
 use eloquentFilter\QueryFilter\ModelFilters\Filterable;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\Pivot;
-use OwenIt\Auditing\Contracts\Auditable;
 use OwenIt\Auditing\Auditable as AuditableModel;
+use OwenIt\Auditing\Contracts\Auditable;
 
+
+/**
+ * @property mixed $firmada
+ * @property Carbon|mixed $fecha_finalizacion
+ * @property mixed $km_final
+ * @property mixed $km_inicial
+ */
 class BitacoraVehicular extends Pivot implements Auditable
 {
     use HasFactory;
@@ -51,7 +58,7 @@ class BitacoraVehicular extends Pivot implements Auditable
 
 
 
-    private static $whiteListFilter = ['*'];
+    private static array $whiteListFilter = ['*'];
     /**
      * ______________________________________________________________________________________
      * RELACIONES CON OTRAS TABLAS
@@ -95,7 +102,7 @@ class BitacoraVehicular extends Pivot implements Auditable
     }
 
     /**
-     * Relación para obtener la ultima notificacion de un modelo dado.
+     * Relación para obtener la última notificacion de un modelo dado.
      */
     public function latestNotificacion()
     {
@@ -111,7 +118,7 @@ class BitacoraVehicular extends Pivot implements Auditable
 
     public static function crearBitacora($data)
     {
-        $bitacora = new BitacoraVehicular([
+        return new BitacoraVehicular([
             'fecha' => $data['fecha'],
             'hora_salida' => $data['hora_salida'],
             'hora_llegada' => $data['hora_llegada'],
@@ -123,7 +130,5 @@ class BitacoraVehicular extends Pivot implements Auditable
             // 'chofer_id' => $data['chofer_id'],
             'vehiculo_id' => $data['vehiculo_id'],
         ]);
-
-        return $bitacora;
     }
 }
