@@ -13,7 +13,7 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('rrhh_solicitudes_puestos_empleos', function (Blueprint $table) {
+        Schema::create('rrhh_contratacion_solicitudes_nuevas_vacantes', function (Blueprint $table) {
             $table->id();
             $table->string('nombre');
             $table->boolean('publicada')->default(false);
@@ -21,16 +21,17 @@ return new class extends Migration
             $table->unsignedBigInteger('autorizador_id');
             $table->unsignedBigInteger('autorizacion_id');
             $table->unsignedBigInteger('cargo_id')->nullable();
+            $table->text('areas_conocimiento')->nullable();
             $table->longText('descripcion');
             $table->string('anios_experiencia')->nullable();
 
             $table->timestamps();
 
             //Laves foraneas
-            $table->foreign('tipo_puesto_id')->references('id')->on('rrhh_tipos_puestos_trabajos')->cascadeOnUpdate()->cascadeOnDelete();
+            $table->foreign('tipo_puesto_id', 'fk_tipo_puesto')->references('id')->on('rrhh_contratacion_tipos_puestos')->cascadeOnUpdate()->cascadeOnDelete();
             $table->foreign('cargo_id')->references('id')->on('cargos')->cascadeOnUpdate()->cascadeOnDelete();
-            $table->foreign('autorizador_id')->references('id')->on('empleados')->cascadeOnUpdate()->cascadeOnDelete();
-            $table->foreign('autorizacion_id')->references('id')->on('autorizaciones')->cascadeOnUpdate()->cascadeOnDelete();
+            $table->foreign('autorizador_id', 'fk_autorizador_id')->references('id')->on('empleados')->cascadeOnUpdate()->cascadeOnDelete();
+            $table->foreign('autorizacion_id', 'fk_autorizacion_id')->references('id')->on('autorizaciones')->cascadeOnUpdate()->cascadeOnDelete();
         });
     }
 
@@ -41,6 +42,6 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('rrhh_solicitudes_puestos_empleos');
+        Schema::dropIfExists('rrhh_contratacion_solicitudes_nuevas_vacantes');
     }
 };
